refactor: create peerConnectRequestSignChangeHandler prototype
