+++ src/core/agent/agent.types.ts
@@ -62,10 +62,6 @@ enum KeriaStatusEventTypes {
   KeriaStatusChanged = "KeriaStatusChanged",
 }
 
-enum PeerConnectSigningEventTypes {
-  PeerConnectSign = "PeerConnectSign",
-}
-
 interface ConnectionStateChangedEvent extends BaseEventEmitter {
   type: typeof ConnectionEventTypes.ConnectionStateChanged;
   payload: {
@@ -91,15 +87,6 @@ interface KeriaStatusChangedEvent extends BaseEventEmitter {
   type: typeof KeriaStatusEventTypes.KeriaStatusChanged;
   payload: {
     isOnline: boolean;
-  };
-}
-
-interface PeerConnectSigningEvent extends BaseEventEmitter {
-  type: typeof PeerConnectSigningEventTypes.PeerConnectSign;
-  payload: {
-    identifier: string;
-    payload: string;
-    approvalCallback: (approvalStatus: boolean) => void;
   };
 }
 
@@ -150,7 +137,6 @@ export {
   AcdcEventTypes,
   NotificationRoute,
   KeriaStatusEventTypes,
-  PeerConnectSigningEventTypes,
 };
 
 export type {
@@ -168,5 +154,4 @@ export type {
   CreateIdentifierResult,
   IdentifierResult,
   KeriaStatusChangedEvent,
-  PeerConnectSigningEvent,
 };
