+++ src/ui/components/AppWrapper/AppWrapper.tsx
@@ -47,6 +47,7 @@ import { useActivityTimer } from "./hooks/useActivityTimer";
 import { setWalletConnectionsCache } from "../../../store/reducers/walletConnectionsCache";
 import { walletConnectionsFix } from "../../__fixtures__/walletConnectionsFix";
 import { PeerConnection } from "../../../core/cardano/walletConnect/peerConnection";
+import { PeerConnectSigningEvent } from "../../../core/cardano/walletConnect/peerConnection.types";
 
 const connectionStateChangedHandler = async (
   event: ConnectionStateChangedEvent,
@@ -113,6 +114,13 @@ const acdcChangeHandler = async (
   }
 };
 
+const peerConnectRequestSignChangeHandler = async (
+  event: PeerConnectSigningEvent,
+  dispatch: ReturnType<typeof useAppDispatch>
+) => {
+  //TODO: Handle logic for the accept/decline sing request
+};
+
 const AppWrapper = (props: { children: ReactNode }) => {
   const dispatch = useAppDispatch();
   const authentication = useAppSelector(getAuthentication);
@@ -150,10 +158,6 @@ const AppWrapper = (props: { children: ReactNode }) => {
         //   Agent.agent.connections.syncKeriaContacts(),
         //   Agent.agent.credentials.syncACDCs(),
         // ]);
-        //TODO: Currently use the first identity as the selected, will change later
-        const selectedAid = (await Agent.agent.identifiers.getIdentifiers())[0]
-          .id;
-        await PeerConnection.peerConnection.start(selectedAid);
       };
       if (!isMessagesHandled && isOnline) {
         handleMessages();
@@ -312,7 +316,7 @@ const AppWrapper = (props: { children: ReactNode }) => {
     });
     PeerConnection.peerConnection.onPeerConnectRequestSignStateChanged(
       async (event) => {
-        //TODO: Handle request and open the popup
+        return peerConnectRequestSignChangeHandler(event, dispatch);
       }
     );
     dispatch(setInitialized(true));
